Some linting

+++ configure_docker.sh
@@ -1,8 +1,9 @@
-#!/bin/bash
+#!/usr/bin/env bash
 # Configuring Docker
 
 os=$(uname -s)
-if [ "$os" = "Linux" ]; then
+
+if [ "$os" = "Linux" ] && command -V apt >/dev/null 2>&1; then
     # https://docs.docker.com/install/linux/docker-ce/ubuntu/
     sudo apt-get remove docker docker-engine docker.io containerd runc
     sudo apt-get install -y apt-transport-https ca-certificates curl gnupg-agent software-properties-common
@@ -13,16 +14,30 @@ if [ "$os" = "Linux" ]; then
 
     # https://docs.docker.com/install/linux/linux-postinstall/
     sudo groupadd docker
-    sudo usermod -aG docker $USER
+    sudo usermod -aG docker "$USER"
     [ -d "$HOME/.docker" ] && sudo chown "$USER":"$USER" "$HOME/.docker" -R && sudo chmod g+rwx "$HOME/.docker" -R
 
     # https://docs.docker.com/compose/install/
-    sudo curl -L "https://github.com/docker/compose/releases/download/1.24.1/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+    sudo curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
     sudo chmod +x /usr/local/bin/docker-compose
+
+elif [ "$os" = "Linux" ] && command -V pacman >/dev/null 2>&1; then
+    sudo pacman -S --noconfirm docker
+
+    # https://docs.docker.com/install/linux/linux-postinstall/
+    sudo groupadd docker
+    sudo usermod -aG docker "$USER"
+    [ -d "$HOME/.docker" ] && sudo chown "$USER":"$USER" "$HOME/.docker" -R && sudo chmod g+rwx "$HOME/.docker" -R
+
+    # docker-compose
+    yay -S docker-compose
+
 elif [ "$os" = "Darwin" ]; then
     ! brew ls --versions python3 >/dev/null 2>&1 && brew install python3
     pip3 install --upgrade pip setuptools
+
 else
     echo "Unsupported OS: $os"
-    return
+    return 1 >/dev/null 2>&1
+    exit 1
 fi
